--- demo2/src/main/java/controller/OrderController.java
@@ -1,5 +1,0 @@
-package controller;
-
-public class OrderController {
-
-}
