--- demo2/src/main/java/controller/SampleController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class SampleController {
-}
