--- demo2/src/main/java/entity/Business.java
@@ -1,34 +1,0 @@
-package entity;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-@Table(name = "business")
-public class Business {
-
-    @Id
-    @Column(name = "business_id")
-    private String businessId;
-
-    @Column(name = "business_name")
-    private String businessName;
-
-    @Column(name = "business_password")
-    private String businessPassword;
-
-    @Column(name = "business_number")
-    private long businessNumber;
-
-    @Column(name = "business_link")
-    private String businessLink;
-
-
-}
