--- demo2/src/main/java/entity/Sample.java
@@ -1,4 +1,0 @@
-package entity;
-
-public class Sample {
-}
