+++ demo2/src/main/java/main/spring/login/demo2/controller/SampleController.java
@@ -1,0 +1,4 @@
+package main.spring.login.demo2.controller;
+
+public class SampleController {
+}
