+++ demo2/src/main/java/main/spring/login/demo2/repository/BusinessRepository.java
@@ -1,0 +1,7 @@
+package main.spring.login.demo2.repository;
+
+import main.spring.login.demo2.entity.Business;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BusinessRepository extends JpaRepository<Business, String> {
+}
