+++ demo2/src/main/java/main/spring/login/demo2/service/BusinessService.java
@@ -1,0 +1,7 @@
+package main.spring.login.demo2.service;
+
+import main.spring.login.demo2.entity.Business;
+
+public interface BusinessService {
+    Business getBusinessById(String businessId);
+}
