+++ demo2/src/main/java/main/spring/login/demo2/service/OrderService.java
@@ -1,0 +1,9 @@
+package main.spring.login.demo2.service;
+
+import main.spring.login.demo2.entity.Order;
+import org.springframework.stereotype.Service;
+
+@Service
+public interface OrderService {
+    Order getOrderById(int orderNumber);
+}
