--- demo2/src/main/java/repository/OrderRepository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public class OrderRepository {
-}
