+++ demo2/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.application.name=demo2
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
-spring.datasource.url=jdbc:mariadb://3.35.197.144:3306
+spring.datasource.url=jdbc:mariadb://3.35.197.144:3306/oms?useSSL=false
 spring.datasource.username=root
 spring.datasource.password=1234
 spring.jpa.show-sql=true
